Implementazione struttura pagina home
Definizione Home Controller e creazione fragments: navbar.html

+++ src/main/java/it/minervhub/model/Utente.java
@@ -15,7 +15,7 @@ public class Utente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int idUtente;
+    private Long idUtente;
 
     @Column(name = "nome", length = 50, nullable = false)
     private String nome;
